fix: quill_to_html when type HashWithIndifferentAccess
* also refactor: schmooze implementation
